Add still more fixups for using TenDRA out of the box with Hybrid7.
- configure magic to stop autoconf from slamming '-g' into CFLAGS
  just because it doesn't explicitly break compiling NULL files
  Thanks to: Andrew Clausen for helping me think around this
- configure magic to make autoconf pick up the right types and
  includes on the host system, since autoconf uses CPPFLAGS to
  check for includes and types, and TenDRA does standards checking
  even when acting as a plain preprocessor, or whatever autoconf
  is making it do
- configure magic to check for <sys/resource.h>, use accordingly.
- we can't _really_ convert directly between data type pointers to
  function pointers, so convert to a faceless intermediary type first,
  namely uintptr_t (as defined by ISO -- should always exist?).


git-svn-id: http://svn.ratbox.org/svnroot/trunk@9913 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ src/modules.c
@@ -257,9 +257,9 @@ int unload_one_module (char *name, int warn)
   if ((modindex = findmodule_byname (name)) == -1) 
     return -1;
 
-  if( (deinitfunc = (void (*)(void))dlsym (modlist[modindex]->address, 
+  if( (deinitfunc = (void (*)(void))(uintptr_t)dlsym (modlist[modindex]->address, 
 				  "_moddeinit")) 
-		  || (deinitfunc = (void (*)(void))dlsym (modlist[modindex]->address, 
+		  || (deinitfunc = (void (*)(void))(uintptr_t)dlsym (modlist[modindex]->address, 
 				  "__moddeinit")))
   {
     deinitfunc ();
@@ -423,9 +423,9 @@ load_a_module (char *path, int warn, int core)
       return -1;
   }
 
-  initfunc = (void (*)(void))dlsym (tmpptr, "_modinit");
+  initfunc = (void (*)(void))(uintptr_t)dlsym (tmpptr, "_modinit");
   if (initfunc == NULL 
-		  && (initfunc = (void (*)(void))dlsym(tmpptr, "__modinit")) == NULL)
+		  && (initfunc = (void (*)(void))(uintptr_t)dlsym(tmpptr, "__modinit")) == NULL)
   {
     sendto_realops_flags(FLAGS_ALL, L_ALL,
                           "Module %s has no _modinit() function",
